Move CrateMetadata impl block to where CrateMetadata is defined
